Refactor: 사용자 input 값 유효성 검사하는 함수 -> 파일로 분리(checkValidation.js)
